Fixing send attachments
