+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/plugin/listeners/SendAttachmentListener.kt
@@ -1,0 +1,12 @@
+package io.getstream.chat.android.client.plugin.listeners
+
+import io.getstream.chat.android.client.models.Message
+
+public interface SendAttachmentListener {
+
+    public suspend fun onAttachmentSendRequest(
+        channelType: String,
+        channelId: String,
+        message: Message,
+    )
+}
