+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/plugin/listeners/SendMessageListener.kt
@@ -25,12 +25,6 @@ import io.getstream.chat.android.client.utils.Result
  */
 public interface SendMessageListener {
 
-    public suspend fun onMessageSendRequest(
-        channelType: String,
-        channelId: String,
-        message: Message,
-    )
-
     /**
      * Side effect to be invoked when the original request is completed with a response.
      *
