+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/plugin/internal/OfflinePlugin.kt
@@ -28,6 +28,7 @@ import io.getstream.chat.android.client.plugin.listeners.EditMessageListener
 import io.getstream.chat.android.client.plugin.listeners.HideChannelListener
 import io.getstream.chat.android.client.plugin.listeners.QueryChannelListener
 import io.getstream.chat.android.client.plugin.listeners.QueryMembersListener
+import io.getstream.chat.android.client.plugin.listeners.SendAttachmentListener
 import io.getstream.chat.android.client.plugin.listeners.SendMessageListener
 import io.getstream.chat.android.client.plugin.listeners.SendReactionListener
 import io.getstream.chat.android.client.plugin.listeners.ShuffleGiphyListener
@@ -65,6 +66,7 @@ internal class OfflinePlugin(
     private val deleteMessageListener: DeleteMessageListener,
     private val shuffleGiphyListener: ShuffleGiphyListener,
     private val sendMessageListener: SendMessageListener,
+    private val sendAttachmentListener: SendAttachmentListener,
     private val queryMembersListener: QueryMembersListener,
     private val createChannelListener: CreateChannelListener,
     private val provideDependency: (KClass<*>) -> Any? = { null },
@@ -80,7 +82,8 @@ internal class OfflinePlugin(
     ShuffleGiphyListener by shuffleGiphyListener,
     SendMessageListener by sendMessageListener,
     QueryMembersListener by queryMembersListener,
-    CreateChannelListener by createChannelListener {
+    CreateChannelListener by createChannelListener,
+    SendAttachmentListener by sendAttachmentListener{
 
     override val interceptors: List<MessageInterceptor> = emptyList()
     override val errorHandlers: List<ErrorHandler> = emptyList()
